REFACTOR
Refactored the code in scoreTest.

Also fixed something I hadn't thought of. The save functions needs to
return true if they save and false if not. They shouldn't be able to
save over values, therefor it needs to return false in this case.
Changed the tests accordingly.

Because of this, I also removed the 0 results from the tests and created
new tests for them.