COVERAGE
Dice: 100%
YahtzeeSet: 100%
Score: 99,31%

Not happy, but I can't come up with a good solution to make score's
coverage any higher. Currently, one part of an IF statement is not
covered (but I know it's covered by my code, which makes it weird) and
the content of a foreach is not covered (which it also is, since one of
the tests would fail otherwise).