COVERAGE
Dice 100%
YahtzeeSet 100%
Score 99,34%
GameLogic 98,96%
GameLoop 100%

Got another case of a weird coverage result, this time in GameLogic. I
have an if statement this time that I know I test all possible
combinations for (It's a (value && value) type, so I test if both are
true or if either one of them are false) and I still get a negative
coverage result for it. I don't get it.

I guess it doesn't matter but I'd like to know what causes it. Anything
under 100% depresses me.