GREEN
Had to change the test a little. Views are extremly hard to get right
without checking. So I checked of course and edited both the view code
and test therafter. In the future I think I'll leave actual formatting
to the view code and just check so that it outputs what it's suppoed to
output.

Also fixed a bug in Program.cs (ironic, huh?) where I gave a new random
object to each of the dice. This resulted in all of them getting the
same number, all the time. Yahtzee Ahoy!