GREEN
Well. Had to remove most of the test since it was impossible to test
that way due to String.Format. This wouldn't really be a fun test anyway
so I hope it suffices that I test how many times Scorelines there are
and how many times UserConsole.WriteLine is called.

Writing test for views is a pain. Please don't.