REFACTOR
Managed to fix a coverage miss in gamelogic by seperating a if statement
into two if statements. Pointless since I already knew it tested both
things but I guess it's good that it now says it's 100% covered.

Also tried to do the same in Score for the CheckForStraight method.
Didn't work there though and I believe it has something to do with the
fact that I'm checking a sorted array for positions of numbers and
compare them to eachother. Apparently codecoverage can't do that (even
though I clearly know that I test cases that both passes and don't). Ah
well.

Will try to solve some more coverage problems in YahtzeeView by making a
private method public and create tests for it.